(split)bug: 27565
Invite.createInvite had dtStamp as int.  As a result mobilesync code was using dtStamp in seconds instead of milliseconds to call this function, which was wrong.

http://bugzilla.zimbra.com/show_bug.cgi?id=27565

Copied from Perforce
 Change: 100741

+++ src/java/com/zimbra/cs/mailbox/calendar/Invite.java
@@ -239,7 +239,7 @@ public class Invite {
             String comment, 
             String location,
             String description,
-            int dtStampOrZero,
+            long dtStampOrZero,
             int sequenceNoOrZero,
             String partStat,
             boolean rsvp,
